fix: projen issue, flatten file structure
