--- src/event-bridge/event-pattern/constants.ts
@@ -1,3 +1,0 @@
-export const OPERATIONS = { STARTS_WITH: "startsWith", INCLUDES: "includes" };
-export const INCLUDES_SEARCH_ELEMENT = "searchElement";
-export const STARTS_WITH_SEARCH_STRING = "searchString";
