--- src/event-bridge/event-pattern/utils.ts
@@ -1,16 +1,0 @@
-import { BinaryOp } from "../../expression";
-
-export const invertBinaryOperator = (op: BinaryOp): BinaryOp => {
-  switch (op) {
-    case "<":
-      return ">";
-    case "<=":
-      return ">=";
-    case ">":
-      return "<";
-    case ">=":
-      return "<=";
-    default:
-      return op;
-  }
-};
