+++ src/event-bridge/index.ts
@@ -1,4 +1,4 @@
-export { EventBus } from "./bus";
+export { EventBus } from "./event-bus";
 export * from "./rule";
 export * from "./types";
 // Do not export `pipe`
