--- src/event-bridge/target-input/index.ts
@@ -1,1 +1,0 @@
-export * from "./synth";
